requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.4
openai==1.3.8
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==11.3.0
aiohttp==3.12.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
numpy==1.24.3
tiktoken==0.5.2
